refactor: refactor parser
