+++ app/routes/privateRoutes/restClientPage/components/parser/Parser.module.css
@@ -1,0 +1,10 @@
+.container {
+  display: flex;
+  align-items: center;
+  gap: var(--spacing-4);
+}
+
+.title {
+  color: var(--color-primary);
+  font-size: var(--font-size-xs);
+}
